backend/requirements/base.txt
# Core Django
Django==5.1.3
django-environ==0.11.2

# Database
psycopg2-binary==2.9.10

# GeoDjango
# Note: PostGIS must be installed separately on the system

# REST Framework
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-gis==1.1
django-filter==24.3
drf-spectacular==0.27.2

# CORS
django-cors-headers==4.5.0

# Image handling
Pillow==11.0.0

# Multilingual support
django-modeltranslation==0.19.8

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.2

# YAML (used for `backend/config/ai.yaml`)
PyYAML==6.0.3

# HTTP client (used for Ollama integration)
httpx==0.28.1
